Move logger to internal package
