+++ libnavui-androidauto/api/current.txt
@@ -5,17 +5,6 @@ package com.mapbox.androidauto {
     field public static final com.mapbox.androidauto.ActiveGuidanceState INSTANCE;
   }
 
-  public final class AndroidAutoLog {
-    method public void logAndroidAuto(String message);
-    method public void logAndroidAutoFailure(String message, Throwable? throwable = null);
-    field public static final com.mapbox.androidauto.AndroidAutoLog INSTANCE;
-  }
-
-  public final class AndroidAutoLogKt {
-    method public static void logAndroidAuto(String message);
-    method public static void logAndroidAutoFailure(String message, Throwable? throwable = null);
-  }
-
   public final class ArrivalState extends com.mapbox.androidauto.CarAppState {
     field public static final com.mapbox.androidauto.ArrivalState INSTANCE;
   }
@@ -104,15 +93,6 @@ package com.mapbox.androidauto.car {
     method public com.mapbox.androidauto.car.MainCarContext getMainCarContext();
     method public suspend Object? observeCarAppState(kotlin.coroutines.Continuation<? super kotlin.Unit>);
     property public final com.mapbox.androidauto.car.MainCarContext mainCarContext;
-  }
-
-  public final class MainScreenManagerKt {
-    method public static void replace(androidx.car.app.ScreenManager, androidx.car.app.Screen screen);
-  }
-
-  public final class RendererUtils {
-    method public int dpToPx(android.content.Context, int dp);
-    field public static final com.mapbox.androidauto.car.RendererUtils INSTANCE;
   }
 
 }
