+++ libnavui-androidauto/src/main/java/com/mapbox/androidauto/car/navigation/CarNavigationInfoObserver.kt
@@ -2,7 +2,7 @@ package com.mapbox.androidauto.car.navigation
 
 import androidx.car.app.navigation.model.NavigationTemplate
 import androidx.car.app.navigation.model.TravelEstimate
-import com.mapbox.androidauto.logAndroidAuto
+import com.mapbox.androidauto.internal.logAndroidAuto
 import com.mapbox.bindgen.Expected
 import com.mapbox.maps.MapboxExperimental
 import com.mapbox.navigation.base.trip.model.RouteProgress
