+++ libnavui-androidauto/src/main/java/com/mapbox/androidauto/car/placeslistonmap/PlaceMarkerRenderer.kt
@@ -10,7 +10,7 @@ import androidx.car.app.model.CarIcon
 import androidx.core.graphics.drawable.IconCompat
 import androidx.vectordrawable.graphics.drawable.VectorDrawableCompat
 import com.mapbox.androidauto.R
-import com.mapbox.androidauto.car.RendererUtils.dpToPx
+import com.mapbox.androidauto.internal.car.RendererUtils.dpToPx
 
 /**
  * Render bitmaps that can be shown as markers on the map.
