+++ libnavui-androidauto/src/main/java/com/mapbox/androidauto/car/search/SearchScreen.kt
@@ -16,8 +16,8 @@ import com.mapbox.androidauto.car.feedback.ui.buildSearchPlacesCarFeedbackProvid
 import com.mapbox.androidauto.car.preview.CarRoutePreviewScreen
 import com.mapbox.androidauto.car.preview.CarRouteRequestCallback
 import com.mapbox.androidauto.car.preview.RoutePreviewCarContext
-import com.mapbox.androidauto.logAndroidAuto
-import com.mapbox.androidauto.logAndroidAutoFailure
+import com.mapbox.androidauto.internal.logAndroidAuto
+import com.mapbox.androidauto.internal.logAndroidAutoFailure
 import com.mapbox.navigation.base.route.NavigationRoute
 import com.mapbox.search.result.SearchSuggestion
