+++ libnavui-androidauto/src/main/java/com/mapbox/androidauto/internal/car/RendererUtils.kt
@@ -1,4 +1,4 @@
-package com.mapbox.androidauto.car
+package com.mapbox.androidauto.internal.car
 
 import android.content.Context
 
@@ -6,7 +6,7 @@ object RendererUtils {
     private const val ANDROID_BASELINE_DPI = 160.0
 
     /**
-     * The car library ignores does not use context.resources.displayMetrics
+     * The car library does not use context.resources.displayMetrics
      * In order to scale correctly on the head unit, use resource.configuration.densityDpi
      */
     fun Context.dpToPx(dp: Int): Int =
