+++ libnavui-androidauto/src/main/java/com/mapbox/androidauto/surfacelayer/textview/CarScene2d.kt
@@ -2,7 +2,7 @@ package com.mapbox.androidauto.surfacelayer.textview
 
 import android.graphics.Rect
 import android.opengl.Matrix
-import com.mapbox.androidauto.logAndroidAuto
+import com.mapbox.androidauto.internal.logAndroidAuto
 import com.mapbox.androidauto.surfacelayer.CarSurfaceLayer
 import com.mapbox.androidauto.surfacelayer.GLUtils
 import com.mapbox.maps.EdgeInsets
